+++ page.php
@@ -14,10 +14,10 @@
 
 									<?php if ( !is_front_page() ) : ?>
 										<h1 class="page-title" itemprop="headline"><?php the_title(); ?></h1>
-									<?php endif; ?>
+									<?php endif; /*?>
 									<p class="byline vcard"><?php
 										printf(__('Posted <time class="updated" datetime="%1$s" pubdate>%2$s</time> by <span class="author">%3$s</span>.', 'bonestheme'), get_the_time('Y-m-j'), get_the_time(__('F jS, Y', 'bonestheme')), bones_get_the_author_posts_link());
-									?></p>
+									?></p><?php */ ?>
 									
 								</header> <!-- end article header -->