+++ CHANGES.md
@@ -1,5 +1,10 @@
 #### 1.1.3
 * CSS fix for sidebars in column-style layout
+* Style improvements for "read more" links, bylines, pagination, and comments
+* Now using an icon font with a smaller filesize than the one used before
+* Removed InfiniteWP from the list of recommended plugins
+* Improved "Output latest posts" widget
+* Added "insert button" button to TinyMCE
 
 #### 1.1.2
 * Numerous CSS improvements to background patterns, the navigation bar, and column-style layout